trnslated and added some reports"

+++ db/migrate/20120306181541_create_participants.rb
@@ -4,7 +4,7 @@ class CreateParticipants < ActiveRecord::Migration
       t.string :name
       t.string :ctg
       t.integer :region
-      t.integer :card
+      t.bigint :card
       t.references :user
 
       t.timestamps
